requirements.txt
# Modal SDK
modal

# Discord Bot
discord.py>=2.3.0
python-dotenv>=1.0.0

# Image Processing & Generation
pillow>=10.0.0
svgwrite>=1.4.3
cairosvg>=2.7.0

# AI/ML Dependencies
torch>=2.4.0
huggingface-hub>=0.36.0

# Modal Logo Generator Dependencies (from stat.sh)
vllm==0.11.2
flashinfer-python==0.5.2
